♻️ refactor: simplify sandbox package layout
