+++ internal/sandbox/boxsh_session_test.go
@@ -1,4 +1,4 @@
-package boxsh
+package sandbox
 
 import (
 	"strings"
