+++ internal/sandbox/local_session.go
@@ -1,4 +1,4 @@
-package local
+package sandbox
 
 import (
 	"bytes"
